Sparse Flux Matrices now use Int32 rather than Int64

+++ src/Fluxes/FluxStructs.jl
@@ -8,18 +8,18 @@ struct FluxMatricesStruct{T<:Union{Float32,Float64}}
     Ap_Flux::Vector{T}                  # Forward Boundary time flux 
     Am_Flux::Vector{T}                  # Backward Boundary time flux
     
-    X_Flux::SparseMatrixCSC{T,Int64}    # sum of space fluxes 
-    P_Flux::SparseMatrixCSC{T,Int64}    # sum of momentum fluxes 
+    X_Flux::SparseMatrixCSC{T,Int32}    # sum of space fluxes 
+    P_Flux::SparseMatrixCSC{T,Int32}    # sum of momentum fluxes 
 
     Vol::Vector{T}                      # SpaceTime volume element
 
-    B_Flux::SparseMatrixCSC{T,Int64}    # B Flux through x boundaries
-    C_Flux::SparseMatrixCSC{T,Int64}    # C Flux through y boundaries
-    D_Flux::SparseMatrixCSC{T,Int64}    # D Flux through z boundaries
+    B_Flux::SparseMatrixCSC{T,Int32}    # B Flux through x boundaries
+    C_Flux::SparseMatrixCSC{T,Int32}    # C Flux through y boundaries
+    D_Flux::SparseMatrixCSC{T,Int32}    # D Flux through z boundaries
 
-    I_Flux::SparseMatrixCSC{T,Int64}    # I Flux through px boundaries
-    J_Flux::SparseMatrixCSC{T,Int64}    # J Flux through py boundaries
-    K_Flux::SparseMatrixCSC{T,Int64}    # K Flux through pz boundaries
+    I_Flux::SparseMatrixCSC{T,Int32}    # I Flux through px boundaries
+    J_Flux::SparseMatrixCSC{T,Int32}    # J Flux through py boundaries
+    K_Flux::SparseMatrixCSC{T,Int32}    # K Flux through pz boundaries
 
 end
